# Conflicts:
#	app/build.gradle
#	app/src/main/java/uz/koinot/stadion/data/model/Dashboard.kt
#	app/src/main/java/uz/koinot/stadion/ui/screens/dashboard/DashboardFragment.kt
#	app/src/main/res/layout/fragment_dashboard.xml

+++ app/src/main/java/uz/koinot/stadion/data/model/Dashboard.kt
@@ -2,5 +2,5 @@ package uz.koinot.stadion.data.model
 
 data class Dashboard(
     var day:String,
-    var benefit:Double,
+    var benefit:Float,
 )
